refactor: use BadRequest err for internal query param errs
