+++ internal/errors.go
@@ -1,0 +1,143 @@
+package internal
+
+import (
+	"fmt"
+	"net/http"
+)
+
+// ErrorWithStatus is an interface that can be implemented by an error to provide
+// a status code
+type ErrorWithStatus interface {
+	error
+	StatusCode() int
+}
+
+// ErrorWithDetail is an interface that can be implemented by an error to provide
+// an additional detail message about the error
+type ErrorWithDetail interface {
+	error
+	DetailMsg() string
+}
+
+// HTTPError is the error response used by the serialization part of the framework.
+type HTTPError struct {
+	// Developer readable error message. Not shown to the user to avoid security leaks.
+	Err error `json:"-" xml:"-" yaml:"-"`
+	// URL of the error type. Can be used to lookup the error in a documentation
+	Type string `json:"type,omitempty" xml:"type,omitempty" yaml:"type,omitempty" description:"URL of the error type. Can be used to lookup the error in a documentation"`
+	// Short title of the error
+	Title string `json:"title,omitempty" xml:"title,omitempty" yaml:"title,omitempty" description:"Short title of the error"`
+	// HTTP status code. If using a different type than [HTTPError], for example [BadRequestError], this will be automatically overridden after Fuego error handling.
+	Status int `json:"status,omitempty" xml:"status,omitempty" yaml:"status,omitempty" description:"HTTP status code" example:"403"`
+	// Human readable error message
+	Detail   string      `json:"detail,omitempty" xml:"detail,omitempty" yaml:"detail,omitempty" description:"Human readable error message"`
+	Instance string      `json:"instance,omitempty" xml:"instance,omitempty" yaml:"instance,omitempty"`
+	Errors   []ErrorItem `json:"errors,omitempty" xml:"errors,omitempty" yaml:"errors,omitempty"`
+}
+
+type ErrorItem struct {
+	More   map[string]any `json:"more,omitempty" xml:"more,omitempty" description:"Additional information about the error"`
+	Name   string         `json:"name" xml:"name" description:"For example, name of the parameter that caused the error"`
+	Reason string         `json:"reason" xml:"reason" description:"Human readable error message"`
+}
+
+func (e HTTPError) Error() string {
+	code := e.StatusCode()
+	title := e.Title
+	if title == "" {
+		title = http.StatusText(code)
+		if title == "" {
+			title = "HTTP Error"
+		}
+	}
+	msg := fmt.Sprintf("%d %s", code, title)
+
+	detail := e.DetailMsg()
+	if detail == "" {
+		return msg
+	}
+
+	return fmt.Sprintf("%s: %s", msg, e.Detail)
+}
+
+func (e HTTPError) StatusCode() int {
+	if e.Status == 0 {
+		return http.StatusInternalServerError
+	}
+	return e.Status
+}
+
+func (e HTTPError) DetailMsg() string {
+	return e.Detail
+}
+
+func (e HTTPError) Unwrap() error { return e.Err }
+
+// BadRequestError is an error used to return a 400 status code.
+type BadRequestError HTTPError
+
+var _ ErrorWithStatus = BadRequestError{}
+
+func (e BadRequestError) Error() string { return e.Err.Error() }
+
+func (e BadRequestError) StatusCode() int { return http.StatusBadRequest }
+
+func (e BadRequestError) Unwrap() error { return HTTPError(e) }
+
+// NotFoundError is an error used to return a 404 status code.
+type NotFoundError HTTPError
+
+var _ ErrorWithStatus = NotFoundError{}
+
+func (e NotFoundError) Error() string { return e.Err.Error() }
+
+func (e NotFoundError) StatusCode() int { return http.StatusNotFound }
+
+func (e NotFoundError) Unwrap() error { return HTTPError(e) }
+
+// UnauthorizedError is an error used to return a 401 status code.
+type UnauthorizedError HTTPError
+
+var _ ErrorWithStatus = UnauthorizedError{}
+
+func (e UnauthorizedError) Error() string { return e.Err.Error() }
+
+func (e UnauthorizedError) StatusCode() int { return http.StatusUnauthorized }
+
+func (e UnauthorizedError) Unwrap() error { return HTTPError(e) }
+
+// ForbiddenError is an error used to return a 403 status code.
+type ForbiddenError HTTPError
+
+var _ ErrorWithStatus = ForbiddenError{}
+
+func (e ForbiddenError) Error() string { return e.Err.Error() }
+
+func (e ForbiddenError) StatusCode() int { return http.StatusForbidden }
+
+func (e ForbiddenError) Unwrap() error { return HTTPError(e) }
+
+// ConflictError is an error used to return a 409 status code.
+type ConflictError HTTPError
+
+var _ ErrorWithStatus = ConflictError{}
+
+func (e ConflictError) Error() string { return e.Err.Error() }
+
+func (e ConflictError) StatusCode() int { return http.StatusConflict }
+
+func (e ConflictError) Unwrap() error { return HTTPError(e) }
+
+// InternalServerError is an error used to return a 500 status code.
+type InternalServerError = HTTPError
+
+// NotAcceptableError is an error used to return a 406 status code.
+type NotAcceptableError HTTPError
+
+var _ ErrorWithStatus = NotAcceptableError{}
+
+func (e NotAcceptableError) Error() string { return e.Err.Error() }
+
+func (e NotAcceptableError) StatusCode() int { return http.StatusNotAcceptable }
+
+func (e NotAcceptableError) Unwrap() error { return HTTPError(e) }
